requirements.txt
Flask==2.0.1
Flask-Login==0.6.2
Werkzeug==2.2.2
tensorflow==2.9.1
tensorflow-hub==0.12.0
numpy==1.21.5
opencv-python==4.6.0.66
Pillow==9.2.0
python-magic==0.4.27
requests==2.28.1
scikit-learn
matplotlib
pandas
face-recognition
dlib
python-dotenv
